Выполин создание единого файла с прогресс баром. todo: 0) добавить файл состояния(что было скачано) 1) удалять скачанные файлы 2) добавить возможность разбивать большие файлы на файлы поменьше(ограниченные по размерам)

+++ main.py
@@ -3,10 +3,11 @@ import pathlib
 import re
 import sys
 from threading import Thread
-from zipfile import ZipFile
+from zipfile import ZipFile, BadZipFile
 
 import requests
 from bs4 import BeautifulSoup
+from rich.pretty import pretty_repr
 from rich.progress import Progress, TaskID
 
 from abstract_info_fetcher import AbstractInfoFetcher, MangaChanInfoFetcher, \
@@ -198,7 +199,7 @@ def extract_num(line: str):
 def filter_exists(folder_prefix: str, chapter_names: list[str], pure_chapter_names: list[str],
                   download_links: list[str]):
     listdir = {file for file in os.listdir(folder_prefix)}
-    listdir = {extract_num(x): os.path.getsize(f'{"/".join(pathlib.Path(folder_prefix).parts)}/{x}') for x in listdir}
+    listdir = {extract_num(x): os.path.getsize(f'{"/".join(pathlib.Path(folder_prefix).parts)}/{x}') for x in listdir if x != 'tmp'}
     remove_chapters = []
     remove_links = []
     remove_pure_chapter_names = []
@@ -271,7 +272,7 @@ def download_manga(folder_prefix: str, progress_bar: Progress, fetcher: Abstract
         download_manga(folder_prefix, progress_bar, fetcher, download_manga_url, title_name)
 
 
-def manga_file_count(folder_prefix):
+def manga_file_count(folder_prefix) -> tuple[list[str], int]:
     files = os.listdir(folder_prefix)
     count = 0
     for f in files:
@@ -279,44 +280,63 @@ def manga_file_count(folder_prefix):
         #     f = f.replace(".cbz", "")
         # elif f.endswith(".cbr"):
         #     f = f.replace(".cbr", "")
-        with ZipFile(os.path.join(folder_prefix, f), "r") as z:
-            namelist = z.namelist()
-            count += len(namelist)
-    return count
+        try:
+            join = os.path.join(folder_prefix, f)
+            with ZipFile(join, "r") as z:
+                namelist = z.namelist()
+                count += len(namelist)
+        except BadZipFile as b:
+            print(f"error in {join}")
+    return files, count
 
 
-def unzip_files(folder_prefix, power, merge_ext: str):
-    files = os.listdir(folder_prefix)
-    files.sort()
-    tmp_dir = os.path.join(folder_prefix, "tmp")
-    os.makedirs(tmp_dir, exist_ok=True)
+def merge(folder_prefix, power, merge_ext: str, progress_bar:Progress, files:list[str], title_name=None, task=None):
     i = 0
-    rez_file = os.path.join(folder_prefix, f'result{merge_ext}')
+    rez_file = os.path.join(folder_prefix, f'{title_name}{merge_ext}')
+    # print(f"Создание результирующего файла: {rez_file}")
     with ZipFile(rez_file, "w") as rez:
+        # print("Файл создан")
         for f in files:
-            with ZipFile(os.path.join(folder_prefix, f), "r") as z:
+            if not f.endswith(".cbz") and not f.endswith(".cbx") and not f.endswith("cbr") and not f.endswith(".rar") and not f.endswith(".zip"):
+                print(f"Пропуск {f}")
+                continue
+            merge_filename = str(os.path.join(folder_prefix, f))
+            # print(f"обрабатывается архив: {merge_filename}")
+            with ZipFile(file=merge_filename, mode="r") as z:
                 for zf in z.namelist():
                     ext = zf[zf.rfind("."):]
                     filename = f'{i:0{power}}{ext}'
-                    filename = os.path.join(tmp_dir, filename)
+                    filename = os.path.join(folder_prefix, filename)
+                    # print(f"Взят файл {filename}")
                     with open(filename, "wb") as tmp_pic:
                         tmp_pic.write(z.read(zf))
+                    # print(f"Разархивирован файл {filename}")
                     rez.write(filename=filename)
+                    # print(f"файл добавлен в архив {merge_filename}")
                     os.remove(path=filename)
+                    # print(f"файл удален {filename}")
                     i += 1
-    os.rmdir(tmp_dir)
+            progress_bar.advance(task, 1)
+    # os.rmdir(tmp_dir)
+    return rez_file
 
 
 def merge_into_archive(folder_prefix: str, progress_bar: Progress, title_name: str, merge_ext: str):
-    count = manga_file_count(folder_prefix)
+    archive_list, page_count = manga_file_count(folder_prefix)
+    archive_list.sort()
     i = 1
     p = 10
-    link_len = count
+    link_len = page_count
     while p < link_len:
         i += 1
         p = p * 10
-
-    unzip_files(folder_prefix=folder_prefix, power=i, merge_ext=merge_ext)
+    task = progress_bar.add_task(description="Архивация в единый файл...", total=len(archive_list))
+    # task = None
+    print("Архивация...")
+    filename = merge(folder_prefix=folder_prefix, power=i, progress_bar=progress_bar, merge_ext=merge_ext, files=archive_list, title_name=title_name, task=task)
+    # archive_list = [os.path.join(folder_prefix, x) for x in archive_list]
+    # for arch in archive_list:
+    #     os.remove(arch)
 
 
 if __name__ == '__main__':
